Update to `Layer 161` + Added New Errors + Some BugFixes (#9)
* Update API schema to Layer 160

* Added some story error

* Renamed `reply_to_msg_id` to `reply_to`

* Added some new error

* SomeFixes with Layer160

Co-Authored-By: Shrimadhav U K <SpEcHiDe@users.noreply.github.com>

* Add missing `caption_entities` parameters and `utils.parse_text_entities calls`

PR (pyrogram#1253)

* Add `full_name` property to `User` and `Chat` objects

PR (pyrogram#1194)

* Version bump to `2.0.107`

* Improvements to automate tasks using pyrogram

When a user has automated reading of chats/channels/groups content or has the message handler active, the links to the channels usually also appear in the following format: "https://t.me/telegram". Right now, we get a 400 error if we want to authenticate using the "https://t.me/telegram" link.

With this improvement in line 70, the user can use either just the public channel nickname or the full link. It makes it easier for the user to use any type of link and he will have no problem accessing it.

With this change, apart from improving the usability of the library, some issues such as "https://github.com/pyrogram/pyrogram/issues/1077" from August 2022 are also fixed.

* Add `MEDIA_FILE_INVALID` to known errors

* Control the number of cached messages via `message_cache` in Client

This commit adds the optional parameter message_cache to Client to control the number of cached messages.

This can significantly help low-end machines control their memory usage.

* Update API scheme on layer 161

* Added some new errors + BugFixes

Co-Authored-By: Shrimadhav U K <SpEcHiDe@users.noreply.github.com>

* Update API scheme on layer 161 (v2)

---------

Co-authored-by: Shrimadhav U K <SpEcHiDe@users.noreply.github.com>
Co-authored-by: barsikus007 <barsikus07@gmail.com>
Co-authored-by: Alisson Lauffer <alissonvitortc@gmail.com>
Co-authored-by: Daniel García Baameiro <dagaba13@gmail.com>
Co-authored-by: KurimuzonAkuma <arturgayauskas@gmail.com>
Co-authored-by: k4rim <49250671+xk4rim@users.noreply.github.com>